COMMERCE-1002 commerce-openapi-admin - refactor package name util -> mapper
