+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/mapper/v1_0/DTOMapper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.openapi.admin.internal.util.v1_0;
+package com.liferay.commerce.openapi.admin.internal.mapper.v1_0;
 
 import com.liferay.commerce.account.model.CommerceAccount;
 import com.liferay.commerce.currency.model.CommerceCurrency;
