+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/util/v1_0/PriceEntryHelper.java
@@ -14,8 +14,8 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.util.v1_0;
 
+import com.liferay.commerce.openapi.admin.internal.mapper.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.ServiceContextHelper;
-import com.liferay.commerce.openapi.admin.internal.util.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.model.v1_0.PriceEntryDTO;
 import com.liferay.commerce.openapi.core.context.Pagination;
 import com.liferay.commerce.openapi.core.model.CollectionDTO;
