+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/util/v1_0/ProductHelper.java
@@ -14,9 +14,9 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.util.v1_0;
 
+import com.liferay.commerce.openapi.admin.internal.mapper.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.DateConfig;
 import com.liferay.commerce.openapi.admin.internal.resource.util.ServiceContextHelper;
-import com.liferay.commerce.openapi.admin.internal.util.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.model.v1_0.ProductDTO;
 import com.liferay.commerce.openapi.core.context.Pagination;
 import com.liferay.commerce.openapi.core.model.CollectionDTO;
