+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/util/v1_0/UserHelper.java
@@ -14,7 +14,7 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.util.v1_0;
 
-import com.liferay.commerce.openapi.admin.internal.util.v1_0.DTOMapper;
+import com.liferay.commerce.openapi.admin.internal.mapper.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.model.v1_0.UserDTO;
 import com.liferay.commerce.openapi.core.context.Pagination;
 import com.liferay.commerce.openapi.core.model.CollectionDTO;
