+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/v1_0/WebSiteResourceImpl.java
@@ -14,7 +14,7 @@
 
 package com.liferay.commerce.openapi.admin.internal.resource.v1_0;
 
-import com.liferay.commerce.openapi.admin.internal.util.v1_0.DTOMapper;
+import com.liferay.commerce.openapi.admin.internal.mapper.v1_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.model.v1_0.WebSiteDTO;
 import com.liferay.commerce.openapi.admin.resource.v1_0.WebSiteResource;
 import com.liferay.commerce.openapi.core.context.Language;
