+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/resource/v2_0/OrderResourceImpl.java
@@ -17,11 +17,11 @@ package com.liferay.commerce.openapi.admin.internal.resource.v2_0;
 import com.liferay.commerce.context.CommerceContext;
 import com.liferay.commerce.model.CommerceAddress;
 import com.liferay.commerce.model.CommerceOrder;
+import com.liferay.commerce.openapi.admin.internal.mapper.v2_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.v2_0.AddressHelper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.v2_0.OrderHelper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.v2_0.OrderItemHelper;
 import com.liferay.commerce.openapi.admin.internal.resource.util.v2_0.OrderNoteHelper;
-import com.liferay.commerce.openapi.admin.internal.util.v2_0.DTOMapper;
 import com.liferay.commerce.openapi.admin.model.v2_0.AddressDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.OrderDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.OrderItemDTO;
